refactor(system-tests): extract review screen/ JSON editor pod tests
Extract pod JSON editor and review screen integration tests.
