+++ system-tests/services/test-pods.js
@@ -39,107 +39,7 @@ describe("Services", function() {
 
       cy.root().getFormGroupInputFor("Command").type(command);
 
-      cy.get("label").contains("JSON Editor").click();
-
-      cy.get("#brace-editor").contents().asJson().should("deep.equal", [
-        {
-          id: `/${Cypress.env("TEST_UUID")}/${serviceName}`,
-          containers: [
-            {
-              name: "container-1",
-              resources: {
-                cpus: 0.1,
-                mem: 10
-              },
-              exec: {
-                command: {
-                  shell: command
-                }
-              }
-            }
-          ],
-          scaling: {
-            kind: "fixed",
-            instances: 1
-          },
-          networks: [
-            {
-              mode: "host"
-            }
-          ],
-          volumes: [],
-          fetch: [],
-          scheduling: {
-            placement: {
-              constraints: []
-            }
-          }
-        }
-      ]);
-
       cy.get("button").contains("Review & Run").click();
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Service ID")
-        .contains(`/${Cypress.env("TEST_UUID")}/${serviceName}`);
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("CPU")
-        .contains("0.1 (0.1 container-1)");
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Memory")
-        .contains("10 MiB (10 MiB container-1)");
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Disk")
-        .contains("Not Supported");
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("GPU")
-        .contains("Not Supported");
-
-      // per container details
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Container Image")
-        .contains("Not Configured");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Force Pull On Launch")
-        .contains("Not Configured");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("CPUs")
-        .contains("0.1");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Memory")
-        .contains("10 MiB");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Command")
-        .contains(command);
 
       cy.get("button").contains("Run Service").click();
 
@@ -250,129 +150,8 @@ describe("Services", function() {
         .type("{backspace}{backspace}{backspace}{backspace}10");
       cy.root().getFormGroupInputFor("Command").type(cmdline);
 
-      // Check JSON view
-      cy.contains("JSON Editor").click();
-
-      // Check contents of the JSON editor
-      cy.get("#brace-editor").contents().asJson().should("deep.equal", [
-        {
-          id: `/${Cypress.env("TEST_UUID")}/${serviceName}`,
-          containers: [
-            {
-              name: "first-container",
-              resources: {
-                cpus: 0.1,
-                mem: 10
-              },
-              image: {
-                id: "nginx",
-                kind: "DOCKER"
-              },
-              exec: {
-                command: {
-                  shell: cmdline
-                }
-              }
-            },
-            {
-              name: "second-container",
-              resources: {
-                cpus: 0.1,
-                mem: 10
-              },
-              exec: {
-                command: {
-                  shell: cmdline
-                }
-              },
-              image: {
-                id: "nginx",
-                kind: "DOCKER"
-              }
-            }
-          ],
-          scaling: {
-            kind: "fixed",
-            instances: 1
-          },
-          networks: [
-            {
-              mode: "host"
-            }
-          ],
-          volumes: [],
-          fetch: [],
-          scheduling: {
-            placement: {
-              constraints: []
-            }
-          }
-        }
-      ]);
-
       // Click Review and Run
       cy.contains("Review & Run").click();
-
-      // Verify the review screen
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Service ID")
-        .contains(`/${Cypress.env("TEST_UUID")}/${serviceName}`);
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("CPU")
-        .contains("0.2 (0.1 first-container, 0.1 second-container)");
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Memory")
-        .contains("20 MiB (10 MiB first-container, 10 MiB second-container)");
-
-      cy
-        .root()
-        .configurationSection("first-container")
-        .configurationMapValue("Container Image")
-        .contains("nginx");
-      cy
-        .root()
-        .configurationSection("first-container")
-        .configurationMapValue("CPUs")
-        .contains("0.1");
-      cy
-        .root()
-        .configurationSection("first-container")
-        .configurationMapValue("Memory")
-        .contains("10 MiB");
-      cy
-        .root()
-        .configurationSection("first-container")
-        .configurationMapValue("Command")
-        .contains(cmdline);
-
-      cy
-        .root()
-        .configurationSection("second-container")
-        .configurationMapValue("Container Image")
-        .contains("nginx");
-      cy
-        .root()
-        .configurationSection("second-container")
-        .configurationMapValue("CPUs")
-        .contains("0.1");
-      cy
-        .root()
-        .configurationSection("second-container")
-        .configurationMapValue("Memory")
-        .contains("10 MiB");
-      cy
-        .root()
-        .configurationSection("second-container")
-        .configurationMapValue("Command")
-        .contains(cmdline);
 
       // Run service
       cy.get("button.button-primary").contains("Run Service").click();
@@ -508,157 +287,7 @@ describe("Services", function() {
         .parents(".form-control-toggle")
         .click();
 
-      cy.get("label").contains("JSON Editor").click();
-
-      cy.get("#brace-editor").contents().asJson().should("deep.equal", [
-        {
-          id: `/${Cypress.env("TEST_UUID")}/${serviceName}`,
-          containers: [
-            {
-              name: "container-1",
-              resources: {
-                cpus: 0.1,
-                mem: 32
-              },
-              endpoints: [
-                {
-                  name: "http",
-                  containerPort: 8080,
-                  hostPort: 0,
-                  protocol: ["tcp"],
-                  labels: {
-                    VIP_0: `/${Cypress.env("TEST_UUID")}/pod-with-service-address:8080`
-                  }
-                }
-              ],
-              image: {
-                id: containerImage,
-                kind: "DOCKER"
-              },
-              exec: {
-                command: {
-                  shell: command
-                }
-              }
-            }
-          ],
-          scaling: {
-            kind: "fixed",
-            instances: 1
-          },
-          networks: [
-            {
-              name: "dcos",
-              mode: "container"
-            }
-          ],
-          volumes: [],
-          fetch: [],
-          scheduling: {
-            placement: {
-              constraints: []
-            }
-          }
-        }
-      ]);
-
       cy.get("button").contains("Review & Run").click();
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Service ID")
-        .contains(`/${Cypress.env("TEST_UUID")}/${serviceName}`);
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Instances")
-        .contains("1");
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("CPU")
-        .contains("0.1 (0.1 container-1)");
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Memory")
-        .contains("32 MiB (32 MiB container-1)");
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Disk")
-        .contains("Not Supported");
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("GPU")
-        .contains("Not Supported");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Container Image")
-        .contains("Not Configured");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Force Pull On Launch")
-        .contains("Not Configured");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("CPUs")
-        .contains("0.1");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Memory")
-        .contains("32 MiB");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Command")
-        .contains(command);
-
-      cy
-        .root()
-        .configurationSection("Network")
-        .configurationMapValue("Network Type")
-        .contains("Container");
-
-      cy
-        .root()
-        .configurationSection("Service Endpoints")
-        .then(function($serviceEndpointsSection) {
-          // Ensure the section itself exists.
-          expect($serviceEndpointsSection.get().length).to.equal(1);
-
-          const $tableCells = $serviceEndpointsSection.find(
-            "tbody tr:visible td"
-          );
-          const cellValues = [
-            "http",
-            "tcp",
-            "8080",
-            `${Cypress.env("TEST_UUID")}${serviceName}.marathon.l4lb.thisdcos.directory:8080`,
-            "container-1",
-            "Edit"
-          ];
-
-          $tableCells.each(function(index) {
-            expect(this.textContent.trim()).to.equal(cellValues[index]);
-          });
-        });
 
       cy.get("button").contains("Run Service").click();
 
@@ -767,150 +396,7 @@ describe("Services", function() {
         .get('input[name="containers.0.artifacts.2.uri"]')
         .type("http://lorempicsum.com/simpsons/600/400/3");
 
-      cy.get("label").contains("JSON Editor").click();
-
-      cy.get("#brace-editor").contents().asJson().should("deep.equal", [
-        {
-          id: `/${Cypress.env("TEST_UUID")}/${serviceName}`,
-          containers: [
-            {
-              name: "container-1",
-              resources: {
-                cpus: 0.1,
-                mem: 10
-              },
-              exec: {
-                command: {
-                  shell: command
-                }
-              },
-              artifacts: [
-                {
-                  uri: "http://lorempicsum.com/simpsons/600/400/1"
-                },
-                {
-                  uri: "http://lorempicsum.com/simpsons/600/400/2"
-                },
-                {
-                  uri: "http://lorempicsum.com/simpsons/600/400/3"
-                }
-              ]
-            }
-          ],
-          scaling: {
-            kind: "fixed",
-            instances: 1
-          },
-          networks: [
-            {
-              mode: "host"
-            }
-          ],
-          volumes: [],
-          fetch: [],
-          scheduling: {
-            placement: {
-              constraints: []
-            }
-          }
-        }
-      ]);
-
       cy.get("button").contains("Review & Run").click();
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Service ID")
-        .contains(`/${Cypress.env("TEST_UUID")}/${serviceName}`);
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Instances")
-        .contains("1");
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("CPU")
-        .contains("0.1 (0.1 container-1)");
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Memory")
-        .contains("10 MiB (10 MiB container-1)");
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Disk")
-        .contains("Not Supported");
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("GPU")
-        .contains("Not Supported");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Container Image")
-        .contains("Not Configured");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Force Pull On Launch")
-        .contains("Not Configured");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("CPUs")
-        .contains("0.1");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Memory")
-        .contains("10 MiB");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Command")
-        .contains(command);
-
-      cy
-        .root()
-        .configurationSection("Container Artifacts")
-        .then(function($containerArtifacts) {
-          // Ensure the section itself exists.
-          expect($containerArtifacts.get().length).to.equal(1);
-
-          const $tableRows = $containerArtifacts.find("tbody tr:visible");
-
-          expect($tableRows.get().length).to.equal(3);
-
-          const cellValues = [
-            ["http://lorempicsum.com/simpsons/600/400/1", "Edit"],
-            ["http://lorempicsum.com/simpsons/600/400/2", "Edit"],
-            ["http://lorempicsum.com/simpsons/600/400/3", "Edit"]
-          ];
-
-          $tableRows.each(function(rowIndex) {
-            const $tableCells = cy.$(this).find("td");
-
-            $tableCells.each(function(cellIndex) {
-              expect(this.textContent.trim()).to.equal(
-                cellValues[rowIndex][cellIndex]
-              );
-            });
-          });
-        });
 
       cy.get("button").contains("Run Service").click();
 
@@ -999,139 +485,7 @@ describe("Services", function() {
 
       cy.root().getFormGroupInputFor("Service Endpoint Name").type("http");
 
-      cy.get("label").contains("JSON Editor").click();
-
-      cy.get("#brace-editor").contents().asJson().should("deep.equal", [
-        {
-          id: `/${Cypress.env("TEST_UUID")}/${serviceName}`,
-          containers: [
-            {
-              name: "container-1",
-              resources: {
-                cpus: 0.1,
-                mem: 32
-              },
-              endpoints: [
-                {
-                  name: "http",
-                  containerPort: 8080,
-                  hostPort: 0,
-                  protocol: ["tcp"]
-                }
-              ],
-              image: {
-                id: "python:3",
-                kind: "DOCKER"
-              },
-              exec: {
-                command: {
-                  shell: "python3 -m http.server 8080"
-                }
-              }
-            }
-          ],
-          scaling: {
-            kind: "fixed",
-            instances: 1
-          },
-          networks: [
-            {
-              name: "dcos",
-              mode: "container"
-            }
-          ],
-          volumes: [],
-          fetch: [],
-          scheduling: {
-            placement: {
-              constraints: []
-            }
-          }
-        }
-      ]);
-
       cy.get("button").contains("Review & Run").click();
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Service ID")
-        .contains(`/${Cypress.env("TEST_UUID")}/${serviceName}`);
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("CPU")
-        .contains("0.1 (0.1 container-1)");
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Memory")
-        .contains("32 MiB (32 MiB container-1)");
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Disk")
-        .contains("Not Supported");
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("GPU")
-        .contains("Not Supported");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Container Image")
-        .contains("python:3");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Force Pull On Launch")
-        .contains("Not Configured");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("CPUs")
-        .contains("0.1");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Memory")
-        .contains("32 MiB");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Command")
-        .contains(command);
-
-      cy
-        .root()
-        .configurationSection("Network")
-        .configurationMapValue("Network Type")
-        .contains("Container");
-
-      cy
-        .root()
-        .configurationSection("Service Endpoints")
-        .then(function($serviceEndpointsSection) {
-          const $tableRow = $serviceEndpointsSection.find("tr:visible");
-          const $tableCells = $tableRow.find("td");
-          const cellValues = ["http", "tcp", "8080", "container-1", "Edit"];
-
-          expect($tableCells.length).to.equal(5);
-
-          $tableCells.each(function(index) {
-            expect(this.textContent.trim()).to.equal(cellValues[index]);
-          });
-        });
 
       cy.get("button").contains("Run Service").click();
 
@@ -1226,138 +580,7 @@ describe("Services", function() {
       cy.root().getFormGroupInputFor("Name").type("test");
       cy.root().getFormGroupInputFor("Container Path").type("test");
 
-      cy.get("#brace-editor").contents().asJson().should("deep.equal", [
-        {
-          id: `/${Cypress.env("TEST_UUID")}/${serviceName}`,
-          containers: [
-            {
-              name: "container-1",
-              resources: {
-                cpus: 0.1,
-                mem: 10
-              },
-              exec: {
-                command: {
-                  shell: command
-                }
-              },
-              volumeMounts: [
-                {
-                  name: "test",
-                  mountPath: "test"
-                }
-              ]
-            }
-          ],
-          scaling: {
-            kind: "fixed",
-            instances: 1
-          },
-          networks: [
-            {
-              mode: "host"
-            }
-          ],
-          volumes: [
-            {
-              name: "test"
-            }
-          ],
-          fetch: [],
-          scheduling: {
-            placement: {
-              constraints: []
-            }
-          }
-        }
-      ]);
-
       cy.get("button").contains("Review & Run").click();
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Service ID")
-        .contains(`/${Cypress.env("TEST_UUID")}/${serviceName}`);
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Instances")
-        .contains("1");
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("CPU")
-        .contains("0.1 (0.1 container-1)");
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Memory")
-        .contains("10 MiB (10 MiB container-1)");
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Disk")
-        .contains("Not Supported");
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("GPU")
-        .contains("Not Supported");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Container Image")
-        .contains("Not Configured");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Force Pull On Launch")
-        .contains("Not Configured");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("CPUs")
-        .contains("0.1");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Memory")
-        .contains("10 MiB");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Command")
-        .contains(command);
-
-      cy.root().configurationSection("Storage").then(function($storageSection) {
-        const $tableRow = $storageSection.find("tbody tr:visible");
-        const $tableCells = $tableRow.find("td");
-        const cellValues = [
-          "test",
-          "EPHEMERAL",
-          "FALSE",
-          "test",
-          "container-1",
-          "Edit"
-        ];
-
-        expect($tableCells.length).to.equal(6);
-
-        $tableCells.each(function(index) {
-          expect(this.textContent.trim()).to.equal(cellValues[index]);
-        });
-      });
 
       cy.get("button").contains("Run Service").click();
 
@@ -1457,144 +680,7 @@ describe("Services", function() {
 
       cy.root().get('input[name="env.3.value"]').type("test");
 
-      cy.get("label").contains("JSON Editor").click();
-
-      cy.get("#brace-editor").contents().asJson().should("deep.equal", [
-        {
-          id: `/${Cypress.env("TEST_UUID")}/${serviceName}`,
-          containers: [
-            {
-              name: "container-1",
-              resources: {
-                cpus: 0.1,
-                mem: 10
-              },
-              exec: {
-                command: {
-                  shell: command
-                }
-              }
-            }
-          ],
-          scaling: {
-            kind: "fixed",
-            instances: 1
-          },
-          networks: [
-            {
-              mode: "host"
-            }
-          ],
-          environment: {
-            camelCase: "test",
-            snake_case: "test",
-            lowercase: "test",
-            UPPERCASE: "test"
-          },
-          volumes: [],
-          fetch: [],
-          scheduling: {
-            placement: {
-              constraints: []
-            }
-          }
-        }
-      ]);
-
       cy.get("button").contains("Review & Run").click();
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Service ID")
-        .contains(`/${Cypress.env("TEST_UUID")}/${serviceName}`);
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Instances")
-        .contains("1");
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("CPU")
-        .contains("0.1 (0.1 container-1)");
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Memory")
-        .contains("10 MiB (10 MiB container-1)");
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Disk")
-        .contains("Not Supported");
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("GPU")
-        .contains("Not Supported");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Container Image")
-        .contains("Not Configured");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Force Pull On Launch")
-        .contains("Not Configured");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("CPUs")
-        .contains("0.1");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Memory")
-        .contains("10 MiB");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Command")
-        .contains(command);
-
-      cy
-        .root()
-        .configurationSection("Environment Variables")
-        .then(function($envSection) {
-          expect($envSection.get().length).to.equal(1);
-
-          const $tableRows = $envSection.find("tbody tr:visible");
-          const cellValues = [
-            ["camelCase", "test", "Shared", "Edit"],
-            ["snake_case", "test", "Shared", "Edit"],
-            ["lowercase", "test", "Shared", "Edit"],
-            ["UPPERCASE", "test", "Shared", "Edit"]
-          ];
-
-          $tableRows.each(function(rowIndex) {
-            const $tableCells = cy.$(this).find("td");
-
-            expect($tableCells.length).to.equal(4);
-
-            $tableCells.each(function(cellIndex) {
-              expect(this.textContent.trim()).to.equal(
-                cellValues[rowIndex][cellIndex]
-              );
-            });
-          });
-        });
 
       cy.get("button").contains("Run Service").click();
 
@@ -1708,145 +794,8 @@ describe("Services", function() {
       cy.get('input[name="labels.3.key"]').type("UPPERCASE");
       cy.get('input[name="labels.3.value"]').type("test");
 
-      // Check JSON view
-      cy.contains("JSON Editor").click();
-
-      // Check contents of the JSON editor
-      cy.get("#brace-editor").contents().asJson().should("deep.equal", [
-        {
-          id: `/${Cypress.env("TEST_UUID")}/${serviceName}`,
-          containers: [
-            {
-              name: containerName,
-              resources: {
-                cpus: 0.1,
-                mem: 10
-              },
-              exec: {
-                command: {
-                  shell: cmdline
-                }
-              }
-            }
-          ],
-          scaling: {
-            kind: "fixed",
-            instances: 1
-          },
-          networks: [
-            {
-              mode: "host"
-            }
-          ],
-          labels: {
-            camelCase: "test",
-            snake_case: "test",
-            lowercase: "test",
-            UPPERCASE: "test"
-          },
-          volumes: [],
-          fetch: [],
-          scheduling: {
-            placement: {
-              constraints: []
-            }
-          }
-        }
-      ]);
-
       // Click Review and Run
       cy.contains("Review & Run").click();
-
-      // Verify the review screen
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Service ID")
-        .contains(`/${Cypress.env("TEST_UUID")}/${serviceName}`);
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Instances")
-        .contains("1");
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("CPU")
-        .contains(`0.1 (0.1 ${containerName})`);
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Memory")
-        .contains(`10 MiB (10 MiB ${containerName})`);
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("Disk")
-        .contains("Not Supported");
-
-      cy
-        .root()
-        .configurationSection("General")
-        .configurationMapValue("GPU")
-        .contains("Not Supported");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Container Image")
-        .contains("Not Configured");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Force Pull On Launch")
-        .contains("Not Configured");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("CPUs")
-        .contains("0.1");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Memory")
-        .contains("10 MiB");
-
-      cy
-        .root()
-        .configurationSection("Containers")
-        .configurationMapValue("Command")
-        .contains(cmdline);
-
-      cy.root().configurationSection("Labels").then(function($section) {
-        expect($section.get().length).to.equal(1);
-
-        const $tableRows = $section.find("tbody tr:visible");
-        const cellValues = [
-          ["camelCase", "test", "Shared", "Edit"],
-          ["snake_case", "test", "Shared", "Edit"],
-          ["lowercase", "test", "Shared", "Edit"],
-          ["UPPERCASE", "test", "Shared", "Edit"]
-        ];
-
-        $tableRows.each(function(rowIndex) {
-          const $tableCells = cy.$(this).find("td");
-
-          expect($tableCells.length).to.equal(4);
-
-          $tableCells.each(function(cellIndex) {
-            expect(this.textContent.trim()).to.equal(
-              cellValues[rowIndex][cellIndex]
-            );
-          });
-        });
-      });
 
       // Run service
       cy.get("button.button-primary").contains("Run Service").click();
